Tests/LibWeb/Text/expected/css/css-imported-image-origin-clean.txt

Imported background image applied: true
Imported image requested: true
Imported image resource timing entries: 0
Imported image initiator type: missing
